+++ pages/_app.tsx
@@ -13,7 +13,7 @@ import posthog from "posthog-js";
 
 import { TipProvider } from "components/atoms/Tooltip/tooltip";
 
-import apiFetcher from "lib/hooks/useSWR";
+import publicApiFetcher from "lib/utils/public-api-fetcher";
 import { initiateAnalytics } from "lib/utils/analytics";
 import { supabase } from "lib/utils/supabase";
 
@@ -96,8 +96,8 @@ function MyApp({ Component, pageProps }: ComponentWithPageLayout) {
         value={{
           revalidateOnFocus: false,
           shouldRetryOnError: false,
-          fetcher: apiFetcher,
-          provider: localStorageProvider
+          fetcher: publicApiFetcher,
+          provider: localStorageProvider,
         }}
       >
         {/* <Toaster position="top-center" /> */}
@@ -112,11 +112,9 @@ function MyApp({ Component, pageProps }: ComponentWithPageLayout) {
               <Component {...pageProps} />
             )}
           </TipProvider>
-          <Script
-            id="ze-snippet"
-            src="https://static.zdassets.com/ekr/snippet.js?key=765edcc9-b888-4651-8b22-79e4365e06f1"
-            strategy="afterInteractive"
-          />
+          <Script id="siteGPT" type="text/javascript">
+            {"d=document;s=d.createElement(\"script\");s.src=\"https://sitegpt.ai/widget/365440930125185604.js\";s.async=1;d.getElementsByTagName(\"head\")[0].appendChild(s);"}
+          </Script>
         </SessionContextProvider>
       </SWRConfig>
     </>